mises à jour des migrations

+++ database/migrations/2024_10_01_000012_add_is_manager_to_employees_table.php
@@ -8,15 +8,19 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('employees', function (Blueprint $table) {
-            $table->boolean('is_manager')->default(false)->after('manager_id');
-        });
+        if (!Schema::hasColumn('employees', 'is_manager')) {
+            Schema::table('employees', function (Blueprint $table) {
+                $table->boolean('is_manager')->default(false)->after('manager_id');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('employees', function (Blueprint $table) {
-            $table->dropColumn('is_manager');
-        });
+        if (Schema::hasColumn('employees', 'is_manager')) {
+            Schema::table('employees', function (Blueprint $table) {
+                $table->dropColumn('is_manager');
+            });
+        }
     }
 };
